+++ src/tooth/toothfile/utils.go
@@ -39,8 +39,8 @@ func parseMetadataPlacement(metadata toothmetadata.Metadata, r *zip.ReadCloser, 
 	return metadata
 }
 
-// getFilePrefix returns the prefix of all files in a zip file.
-func getFilePrefix(r *zip.ReadCloser) string {
+// GetFilePrefix returns the prefix of all files in a zip file.
+func GetFilePrefix(r *zip.ReadCloser) string {
 	prefix := ""
 	for i, file := range r.File {
 		if strings.HasSuffix(file.Name, "/") { // Skip directories.