+++ CHANGELOG.md
@@ -7,14 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unrelease]
 
-## [0.2.1] - 2023-01-11
+## [0.3.0] - 2023-01-20
+
+### Added
+
+- Possession keeping support when force-reinstalling or upgrading.
+- `--force-reinstall` flag and `--upgrade` flag support.
+
+## [0.2.1] - 2023-01-18
 
 ### Fixed
 
 - Failing to fetch tooth whose version has suffix `+incompatible`.
 - Failing to parse wildcards.
 
-## [0.2.0] - 2023-01-11
+## [0.2.0] - 2023-01-18
 
 ### Added
 
@@ -36,7 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/LiteLDev/Lip/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.1.0